refactor (puconvocation-com): Refactor imports for the AttendeeTable component to use a Barrel files scheme.
