+++ frontend/websites/puconvocation-com/src/components/attendee/index.ts
@@ -14,3 +14,4 @@
 export {default as SeatMap} from "./seat_map"
 export {default as Ticket} from "./ticket"
 export {default as VenueMap} from "./venue_map"
+export {default as AttendeeTable} from "./attendee_table"
